requirements.txt
# package required details for python are kept in the requirement.txt
requests
python-dotenv